moving competition container into components folder
